pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fl_fashion_mnist"
version = "1.0.0"
description = ""
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.18.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch>=2.6.1",
    "torchvision==0.22.1",
    "wandb",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "Abenezer"

[tool.flwr.app.components]
serverapp = "fl_fashion_mnist.server_app:app"
clientapp = "fl_fashion_mnist.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-fit = 0.5
local-epochs = 1

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10
